+++ changelog.md
@@ -9,14 +9,14 @@ All notable changes to this project will be documented in this file.
   - `assertArraysAreSame(array $expected, array $actual, ?string $message = null)`
   - `assertArraysAreSameIgnoringOrder(array $expected, array $actual, ?string $message = null)`
   - `assertArraysAreSameIgnoringOrderRecursive(array $expected, array $actual, ?string $message = null)`
-  - `assertStringIsSameAsFile(string $expected, string $actual, ?string $message = null)`
 - trait `WithVFS` that provides adapted interface for `mikey179/vfsstream` with following methods:
   - `addFileToVFS(string $filename, ?int $permissions = null, ?string $content = null)`
   - `addDirectoryToVFS(string $dirname, ?int $permissions = null)`
   - `addStructure(array $structure)`
   - `getFullPath(string $path)`
   - `getEscapedFullPath(string $path)`
-- test data loader from JSON files that wraps each test case into test data container
+- trait `WithDataLoader` that provides method `loadDataFromFile(string $filename)` for easy use `DataLoader` in data
+  providers
 - command `generate-ide-test-containers` that generate service file for IDE used to autocomplete